% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatransform_tplot.R
\name{countheatmapplot}
\alias{countheatmapplot}
\title{Count Heatmap Plot of Specific Genes}
\usage{
countheatmapplot(tdds)
}
\arguments{
\item{tdds}{DESeqTransform Input}
}
\value{
A customised count heatmap plot which is saved into a .png file
}
\description{
Produces a customised heatmap of transformed/normalised counts of genes of your choosing through a series of prompts
}
\examples{
library(DESeq2)
dds <- DESeq2::DESeq(exampledeseq)
tdds <- DESeq2::vst(dds)
tdds <- DESeq2::rlog(dds)
tdds <- DESeq2::normTransform(dds, f = log2, pc = pcount)

library(easybioinfo)
dds <- easybioinfo::rundeseq(expressiondf, md)
tdds <- easybioinfo::datatransform(dds)

countheatmapplot(dds)
}
