% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countplot.R
\name{countplot}
\alias{countplot}
\title{Normalised Count Plots for a Single Gene}
\usage{
countplot(dds)
}
\arguments{
\item{dds}{DESeqDataSet Input}
}
\value{
Count plots of your DESeq2 results in individual .png files
}
\description{
Produces count plots of your DESeq2 results
}
\examples{
expression = easybioinfo::deseqexpr
md = easybioinfo::deseqmd

exampledeseq = DESeq2::DESeqDataSetFromMatrix(countData = expression, colData = md, design = ~condition, tidy = TRUE)
dds <- DESeq2::DESeq(exampledeseq)

library(easybioinfo)
dds <- easybioinfo::rundeseq(expression, md)
countplot(dds)
}
