% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplanmeier.R
\name{customisesurvplot}
\alias{customisesurvplot}
\title{Customise Your Kaplan-Meier Survival Plot}
\usage{
customisesurvplot(unidf)
}
\arguments{
\item{unidf}{A list that is produced from the uniKM function}
}
\value{
Customised Kaplan-Meier survival plots that are saved into .png files
}
\description{
Customise your survival plots based on a series of prompts. Able to customise the font size of the legends, p-values, axis titles, and the title
Also able to customise the position of the p-value and legend
}
\examples{
df <- easybioinfo::kmexpr
md <- easybioinfo::kmclinical
kmdf <- tidykmdata(df, md)
unidf <- uniKM(kmdf)
customisesurvivalplot(unidf)
}
