% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatransform_tplot.R
\name{datatransform}
\alias{datatransform}
\title{Transforms and Normalises DESeq2 Counts for Downstream Analysis}
\usage{
datatransform(dds)
}
\arguments{
\item{dds}{DESeqDataSet Input}
}
\value{
A DESeqTransform object of transformed, normalised counts and saves them into excel files
}
\description{
Transforms your DESeq2 counts via:
\enumerate{
\item Variance Stabilising Transformation
\item Regularised Log Transformation
\item Normalised Count Transformation
}

Please check the DESeq2 vignette for further details on the transformations: http://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html
}
\examples{
expression = easybioinfo::deseqexpr
md = easybioinfo::deseqmd

exampledeseq = DESeq2::DESeqDataSetFromMatrix(countData = expression, colData = md, design = ~condition, tidy = TRUE)
dds <- DESeq2::DESeq(exampledeseq)

library(easybioinfo)
dds <- easybioinfo::rundeseq(expressiondf, md)
datatransform(dds)
}
