% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes2hgnc.R
\name{ddsgenes2hgnc}
\alias{ddsgenes2hgnc}
\alias{ddsgenestohgnc}
\title{Gene Conversion of DESeq2 Results}
\usage{
ddsgenestohgnc(dds, srcType = "ensembl_gene_id")
}
\arguments{
\item{dds}{A DESeqDataSet input}

\item{srcType}{Gene Identifier required by biomaRt. Defaults to "ensembl_gene_id". Can be changed in the function through user prompts}
}
\value{
A DESeqDataSet with gene symbols converted into official HGNC symbols
}
\description{
Converts your DESeq2 result into official HGNC symbols
Conversion is done using the 'biomaRt' package.
}
\examples{
library(DESeq2)
expression = easybioinfo::deseqexpr
md = easybioinfo::deseqmd

exampledeseq = DESeq2::DESeqDataSetFromMatrix(countData = expression, colData = md, design = ~condition, tidy = TRUE)
dds <- DESeq2::DESeq(exampledeseq)

library(easybioinfo)
dds <- easybioinfo::rundeseq(expressiondf, md)
ddsgenestohgnc(dds)
}
