% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easybioinfo-package.R
\docType{package}
\name{easybioinfo-package}
\alias{easybioinfo}
\alias{easybioinfo-package}
\title{easybioinfo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/drgellerbing/easybioinfo}
  \item Report bugs at \url{https://github.com/drgellerbing/easybioinfo/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
