% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes2hgnc.R
\name{genes2hgnc}
\alias{genes2hgnc}
\title{Converts Gene Identifiers to HGNC Symbols}
\usage{
genes2hgnc(expr, srcType = "ensembl_gene_id")
}
\arguments{
\item{expr}{Your expression dataframe}

\item{srcType}{Gene Identifier required by biomaRt. Defaults to "ensembl_gene_id". Can be changed in the function through user prompts}
}
\value{
A dataframe with converted gene identifiers
}
\description{
Using this function, you can easily convert different gene identifiers from your expression dataframe into official HGNC symbols. Conversion is done using the 'biomaRt' package.
}
\examples{
df <- easybioinfo::kmexpr
df <- genes2hgnc(df) 
}
