% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getresults.R
\name{getresults}
\alias{getresults}
\title{DESeq2 Results Processing}
\usage{
getresults(deseq_dds)
}
\arguments{
\item{deseq_dds}{A DESeqDataSet input}
}
\value{
A list containing the DESeqResults, the full results and the significant results. Full results and significant results can be saved into excel files
}
\description{
Processes all your DESeq2 results through a series of prompts. Filters significant results according to your desired p-value and log2 fold change and saves the results into excel files.
}
\examples{
library(DESeq2)
expression = easybioinfo::deseqexpr
md = easybioinfo::deseqmd

exampledeseq = DESeq2::DESeqDataSetFromMatrix(countData = expression, colData = md, design = ~condition, tidy = TRUE)
dds <- DESeq2::DESeq(exampledeseq)

library(easybioinfo)
dds <- easybioinfo::rundeseq(expressiondf, md)
getresults(dds)
}
