% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot.R
\name{maplot}
\alias{maplot}
\title{Customised MA-Plot}
\usage{
maplot(result)
}
\arguments{
\item{result}{Either a DESeqResults object or the list produced from the getresults/shrinklfc function}
}
\value{
A MA plot that can be saved into .png files
}
\description{
A function to produce customised scatter plot of log2 fold changes (on the y-axis) versus the mean of normalised counts (on the x-axis)
}
\examples{
expression = easybioinfo::deseqexpr
md = easybioinfo::deseqmd

exampledeseq = DESeq2::DESeqDataSetFromMatrix(countData = expression, colData = md, design = ~condition, tidy = TRUE)
dds <- DESeq2::DESeq(exampledeseq)
result <- DESeq2::results(dds)
maplot(result)

df <- easybioinfo::deseqexpr
md <- easybioinfo::deseqmd
dds <- rundeseq(df, md)
result <- getresults(dds)
maplot(result)

lfc <- shrinklfc(dds)
maplot(lfc)

}
