% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicox.R
\name{multicox}
\alias{multicox}
\title{Multivariate Cox Proprotional Hazard Analysis}
\usage{
multicox(kmdf)
}
\arguments{
\item{kmdf}{Ideally, dataframe produced from the tidykmdata function. Dataframe should contain "os"/"pfs" columns for OS/PFS values, "statusos"/"statuspfs" with "1 or 0" as events, as well as expression values for individual genes in each column.}
}
\value{
A list containing the class coxph representing the fit, the full cox analysis result, the significant cox analysis result and the transformed input dataframe used for the analysis.
Also produces a forest plot that is saved into a .png file. Forest plot can be further customised using plotforest function.
}
\description{
Performs multivariate cox proportional hazard analysis for overall survival or progression-free survival through a series of prompts.
The function will prompt you for a list of genes.
Survival can be calculated based on four different cutoffs: median, z-scores, quartiles, custom quartiles or optimal.
Optimal cutoffs are calculated by performing a test of independence of response using maximally selected rank statistics
}
\examples{
df <- easybioinfo::kmexpr
md <- easybioinfo::kmclinical
kmdf <- tidykmdata(df, md)
unidf <- uniKM(kmdf)
cox <- multicox(kmdf)
}
