% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatransform_tplot.R
\name{pcaplot}
\alias{pcaplot}
\title{PCA Plotting Function}
\usage{
pcaplot(tdds)
}
\arguments{
\item{tdds}{DESeqTransform Input}
}
\value{
A customised PCA plot which is saved into a .png file
}
\description{
Produces a PCA plot based on your DESeqTransform dataset. Plot parameters can be tweaked specifically by the user
}
\examples{
library(DESeq2)
dds <- DESeq2::DESeq(exampledeseq)
tdds <- vst(dds)
tdds <- rlog(dds)
tdds <- normTransform(dds, f = log2, pc = pcount)

library(easybioinfo)
dds <- easybioinfo::rundeseq(expressiondf, md)
tdds <- easybioinfo::datatransform(dds)

pcaplot(dds)
}
