% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicox.R
\name{plotforest}
\alias{plotforest}
\title{Customise Forest Plot}
\usage{
plotforest(cox)
}
\arguments{
\item{cox}{The list from the multicox function}
}
\value{
A forest plot that can be saved into a .png file
}
\description{
A helper function to allow to produce customised forest plots.
The title, font size, reference label, and the number of digits for estimates and p-values in the plot can all be customised.
}
\examples{
df <- easybioinfo::kmexpr
md <- easybioinfo::kmclinical
kmdf <- tidykmdata(df, md)
unidf <- uniKM(kmdf)
cox <- multicox(kmdf)
plotforest(cox)
}
