% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundeseq.R
\name{rundeseq}
\alias{rundeseq}
\title{Differential Expression Analysis Based on the Gamma-Poisson Distribution}
\usage{
rundeseq(expression, md)
}
\arguments{
\item{expression}{Gene expression dataframe}

\item{md}{Clinical data dataframe}
}
\value{
a DESeqDataSet object with results stored as metadata columns. These results can be accessed by calling results function from "DESeq2" or from the getresults function
}
\description{
A streamlined pipeline that performs differential expression analysis through a series of prompts using the "DESeq2" package.
Produces DESeqDataSet with DESeq2::DESeqDataSetFromMatrix, establishes the control through your prompts, and runs the differential expression analysis using DESeq2::DESeq
}
\examples{
df <- easybioinfo::deseqexpr
md <- easybioinfo::deseqmd
dds <- rundeseq(df, md)
}
