% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{tidyddsmd}
\alias{tidyddsmd}
\title{Preparing Large Metadatas for Differential Expression Analysis}
\usage{
tidyddsmd(exprdf)
}
\arguments{
\item{exprdf}{An expression dataframe containing RNA-Seq counts}
}
\value{
A metadata dataframe for downstream DESeq2 differential expression analysis
}
\description{
This function creates a metadata table required for the differential expression analysis.
This function runs on the basis of common identifiers that R can search for for each experimental condition.
}
\examples{
example_df <- data.frame(id = c("TSPAN6", "TNMD", "DPM1", "SCYL3"),
                         c1 = c(479, 326, 792, 502),
                         c2 = c(325, 12, 1148, 651),
                         c3 = c(286, 10, 748, 233),
                         c4 = c(422, 4, 614, 336),
                         c5 = c(327, 12, 536, 383),
                         c6 = c(123, 42, 452, 21))
colnames(example_df)[2:ncol(example_df)] <- c("GTEX-WZTO", "GTEX-PVOW", "TCGA-13NYS", "TCGA-11NUK", "LGG-11DXW", "LGG-17WES")
The example dataframe has one experimental control and two experimental conditions, GBM and LGG

The metadata table produced from this function for the example would be as below:
example_md <- data.frame(id = c("GTEX-WZTO", "GTEX-PVOW", "TCGA-13NYS", "TCGA-11NUK", "LGG-11DXW", "LGG-17WES"),
                          condition = c("Control", "Control", "GBM", "GBM", "LGG", "LGG"))

To produce the file, you are required to provide unique identifier for each of your experimental condition.
Referring to the example, the unique identifier for the control samples would be 'GTEX'.
Similarly, the unique identifier for the GBM samples would be 'TCGA'.
Lastly, the unique identifier for the LGG sample would be 'LGG'.
}
