% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{tidykmdata}
\alias{tidykmdata}
\title{Data Wrangling for Downstream Survival Analysis}
\usage{
tidykmdata(expressiondataframe, clinicaldataframe)
}
\arguments{
\item{expressiondataframe}{Dataframe containing expression values. The dataframe should have individual genes in each row and individual samples in each column.}

\item{clinicaldataframe}{Dataframe containing clinical data. The clinical data should contain the samples, overall survival/progression-free survival values, as well as the events for either the overall survival/progression-free survival.}
}
\value{
A dataframe containing information for the samples, the survival information and the expression of your query genes for downstream survival analysis using uniKM or multicox functions
}
\description{
A prompt-based function that produces a dataframe containing the overall survival or progression-free survival information for downstream survival analysis.
The function will prompt you for a list of genes and wrangle your expression dataframes and clinical data into one final dataframe.
}
\examples{
df <- easybioinfo::kmexpr
md <- easybioinfo::kmclinical

kmdf <- tidykmdata(df, md)
}
