% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidydata.R
\name{transformexpr}
\alias{transformexpr}
\title{Transforms RNA-Seq Expression into Raw Counts}
\usage{
transformexpr(exprdf)
}
\arguments{
\item{exprdf}{An expression dataframe containing RNA-Seq log2 transformed counts}
}
\value{
A dataframe with transformed raw counts for downstream DESeq2 differential expression analysis.
}
\description{
A function to convert log2 transformed RNA-Seq counts into raw counts for downstream DESeq2 differential expression analysis.
This function will also account for any psudocounts based on your input prompt.
}
\examples{
df <- transformexpr(expr)
}
