% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplanmeier.R
\name{uniKM}
\alias{uniKM}
\title{Kaplan-Meier Survival Plots}
\usage{
uniKM(kmdf)
}
\arguments{
\item{kmdf}{Ideally, dataframe produced from the tidykmdata function. Dataframe should contain "os"/"pfs" columns for OS/PFS values, "statusos"/"statuspfs" with "1 or 0" as events, as well as expression values for individual genes in each column.}
}
\value{
A list containing the complete dataframe with survivals calculated using Cox-Proportional Hazard, a dataframe with only significant results, and the transformed dataframe.
Also produces customised individual Kaplan Meier plots that can be saved into .png files.
The Kaplan-Meier plots can be further customised using the customisesurvplot function
}
\description{
A function that produces customised survival plots based on a series of prompts. Able to produce overall survival (OS) plots or progression-free survival (PFS) plots.
The function will prompt you for a list of genes.
This function also provides the option to plot Schoenfield plots that checks for the proportional hazard assumption.
Survival can be calculated based on four different cutoffs: median, z-scores, quartiles, custom quartiles or optimal.
Optimal cutoffs are calculated by performing a test of independence of response using maximally selected rank statistics
}
\examples{
df <- easybioinfo::kmexpr
md <- easybioinfo::kmclinical
kmdf <- tidykmdata(df, md)
unidf <- uniKM(kmdf)
}
