% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Customised Volcano Plots}
\usage{
volcanoplot(result)
}
\arguments{
\item{result}{Either a DESeqResults object or the list produced from the getresults/shrinklfc function}
}
\value{
Volcano plots of your DESeq2 results that are saved in individual .png files
}
\description{
A function to produce a customised volcano plots. The volcano plot will be plotted based on your desired p-value and LFC prompts
Able to label specific genes on the plot if desired, otherwise, the function will default to labelling all the genes.
Parameters such as the point size, label size, and the axis limits can be customised.
}
\examples{
expression = easybioinfo::deseqexpr
md = easybioinfo::deseqmd

exampledeseq = DESeq2::DESeqDataSetFromMatrix(countData = expression, colData = md, design = ~condition, tidy = TRUE)
dds <- DESeq2::DESeq(exampledeseq)
result <- DESeq2::results(dds)
volcanoplot(result)

lfc <- DESeq2::lfcShrink(dds, coef = 2, type = "apeglm", alpha = 0.05)
volcanoplot(lfc)

library(easybioinfo)
dds <- easybioinfo::rundeseq(expression, md)
results <- getresults(dds)
volcanoplot(results)

lfc <- shrinklfc(dds)
volcanoplot(lfc)
}
